feat: implementação da tela de status do pedido
