--- app/(website)/pedidoconfirmado/page.js
@@ -1,24 +1,0 @@
-
-import React from 'react';
-
-import CardConfirmacao from '@/components/pedidoconfirmado/CardConfirmacao';
-
-
-export default async function PedidoConfirmado(params) {
-
-
-	const pedido = await prisma.order.findFirst({ where: { id: params.id }, include: {user: true, address:true, order_items: {
-		include: {
-		  product: true, 
-		},
-	  },},});
-
-	
-
-	return (
-		<main>
-			{
-                <CardConfirmacao pedido={pedido}></CardConfirmacao>
-                }
-                </main>)}
-
